[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2024 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -4,8 +4,8 @@ using namespace std;
 
 int hashFn(string word) {
     int r = 31;
-    int r_pow = 1;
-    int sum = 0;
+    long long r_pow = 1;
+    long long sum = 0;
 
     for(int i = 0; i < word.length(); i++) {
         int tmp = word[i] - 'a' + 1;
